+++ src/views/MangaWithReferenceView.vue
@@ -1,17 +1,5 @@
 <script setup lang="ts">
-import { ElPopover, ElButton } from 'element-plus'
-import { ZoomIn } from '@element-plus/icons-vue'
-
-const documents: { id: string; title: string }[] = [
-  {
-    id: '1',
-    title: 'example title1'
-  },
-  {
-    id: '2',
-    title: 'example title2'
-  }
-]
+import ReferencePopover from '@/components/mangaWithReference/ReferencePopover.vue'
 </script>
 
 <template>
@@ -20,20 +8,9 @@ const documents: { id: string; title: string }[] = [
     <div :class="$style.images">
       <img src="@/assets/left.jpg" alt="left page" width="400px" :class="$style.image" />
       <img src="@/assets/right.jpg" alt="right page" witdh="400px" :class="$style.image" />
-      <div :class="$style.popover">
-        <el-popover :width="400" trigger="hover" placement="top-start">
-          <template #reference>
-            <el-button type="info" :icon="ZoomIn" circle :class="$style.infoButton" />
-          </template>
-          <ul>
-            <li v-for="document in documents" :key="document.id">
-              <router-link :to="`/documents/documentId=${document.id}`">
-                {{ document.title }}
-              </router-link>
-            </li>
-          </ul>
-        </el-popover>
-      </div>
+      <reference-popover :class="$style.popover1" />
+      <reference-popover :class="$style.popover2" />
+      <reference-popover :class="$style.popover3" />
     </div>
   </div>
 </template>
@@ -46,10 +23,17 @@ const documents: { id: string; title: string }[] = [
 .image {
   width: 400px;
 }
-.popover {
-  position: absolute;
+.popover1 {
   top: 12px;
   left: 600px;
+}
+.popover2 {
+  top: 140px;
+  left: 980px;
+}
+.popover3 {
+  top: 280px;
+  left: 660px;
 }
 .infoButton {
   cursor: zoom-in;
